Added functions to report version information.
git revision and tags are gathered at compile time.
The included cmake modules will force a re-configure (a rerun of cmake)
before a build every time the HEAD commit changes. This means that we
don't need to remember to rerun cmake manually to fix the revision
number.
This also means that version.c is compiled everytime.

New functions:
- rapp_get_version: returns the current version, i.e. 0.1
- rapp_get_version_sha1: returns the git sha1 of the commit
- rapp_get_version_tag : returns the git tag (if any)
- rapp_get_version_full: returns $version.$revision