add the default container and port to new API.
a `default` container is a container which runs independently
from the route (aka by default, hence the name).
If and only if the container doesn't serve - currently implemented
as checking for not-success in container invocation, the usual
search by route, first fit, takes place.

default container is optional and not set by default.
If there isn't a default container, the routing is purely done
by route match, first fit.

a default route can be useful, for example, in the following
scenario:
- you have a bunch of static files in $DOCROOT
- AND your app is mounted on $DOCROOT

without the default container, you'll have, either
- to mount the app under $DOCROOT/something (route = /something)
  and mount a static file container under $DOCROOT/ (route = /)
- to play wih rewrite_rules

or others similar nasty workarounds.
Of course you still can do that or whatever tricks you want, but
at least now you have a choice.